refactor(api): extract rankings from apidata.js, delete the two shims
First pass on #147. apidata.js 1702 -> 1369 lines.

This is a MOVE, not a rewrite. The extracted code is unchanged, and apidata.js
re-exports it, so no call site changes for the moved functions.

api/rankings.js takes buildTierResolver, fetch_global_performance_rankings,
fetch_country_node_counts and _extract_country_counts, plus the
globalPerfRankings_v5 cache and its stale-key pruning.

fetch_country_node_counts had to move too, and the reason is the interesting
part. The build -- not the tests -- failed after the first cut with
GLOBAL_RANKINGS_CACHE_KEY undefined: that function reads the rankings cache
DIRECTLY, opportunistically reusing the full payload when the gamification tab
has already populated it. Leaving it behind would have meant exporting a cache
key and TTL purely so another module could reach into this one's storage, which
is not a boundary worth having. The issue's own proposed split put both in
api/rankings.js; the coupling is why.

api/endpoints.js holds the one explorer path both modules need. Duplicating a
URL across modules is exactly the drift #147 is about -- it is how the same
465 KB endpoint ended up fetched twice inside one Promise.all.

Both shims are gone. home/apidata.js and main/apidata.js were two-line
`export * from '../apidata'` files that made the tree look like it held three
API layers when it held one. Nine importers repointed at 'apidata'. One of them
(demo/Demo.jsx) used a third path form, '../main/apidata', which the first
search missed -- caught by grepping for the bare substring afterwards rather
than trusting the import-style patterns.

Behaviour-preservation is SHOWN, not assumed, which is what the issue asked
for. Beyond 530 passing tests and a clean build, the Track 3 audit harness was
re-run against live data after the split: D1 earnings still agree with the
independent reference on all twelve figures within 1e-9. Tests alone would not
have proved that -- they share the code's assumptions; the reference does not.

Worth noting for the next pass: the full build caught a break that 530 green
tests did not, for the second time this session. A refactor of this shape needs
both.

Remaining in apidata.js for later passes: global stats, node transformation,
parallel assets, app specs, health, GPU pricing, DOS state. Currency was
already extracted to currency.js before this pass.
