--- client/src/home/apidata.js
@@ -1,2 +1,0 @@
-// Shared API module — see client/src/apidata.js
-export * from '../apidata';
